Implemented Dirty State for Maps,
refactored unused imports and unit tests
