+++ src/mapeditor/mod.rs
@@ -1,10 +1,7 @@
 use bevy::prelude::*;
 use bevy::render::camera::RenderTarget;
 use libexodus::tiles::Tile;
-use libexodus::world::GameWorld;
-use crate::AppState;
 use crate::game::tilewrapper::MapWrapper;
-use crate::game::world::WorldPlugin;
 use crate::mapeditor::edit_world::EditWorldPlugin;
 use crate::mapeditor::mapeditor_ui::MapEditorUiPlugin;
 use crate::mapeditor::preview_tile::MapEditorPreviewTilePlugin;
@@ -13,7 +10,6 @@ mod mapeditor_ui;
 mod player_spawn;
 mod preview_tile;
 mod edit_world;
-mod save_file_dialog;
 
 pub struct SelectedTile {
     pub tile: Tile,
